backend/requirements.txt
django
python-decouple
mysqlclient
cryptography
djangorestframework
django-cors-headers